Add ArgoCD and HAProxy operator
This gets us some useful tools as options.

Note: I'm not a huge fan of haproxy operator, you have to apply their
recommended custom resources with validation turned off, however it does
function. I'm open to discussion on this point.

Note: Arguably not the proper place to add argocd, but here we are.