Add pieces (#13)
Add Tailscale operator so we can expose the controlplane API in a
high-availability manner, so that no matter which controlplane fails we
will not have a SPOF.

Add HAProxy operator so that we have a single service / ClusterIP which
load-balances across the controlplane servers that can be exposed via
Tailscale.

Add ArgoCD so that we can test some services (mariaDB operator, forgejo)
as we go.

Significant updates to the README to describe how to use this.