Update to use the default kubernetes service
Occasionally as I work my way through a system, I mentally mark
something as default/core/important, and not to be messed with until I
understand it. I then go through a long process of learning, at the end
of which I know what it is, what it does, and why I should change it.

Usually in a flash of insight at the end.

Which is where I am today,